requirements.txt
beautifulsoup4==4.9.3
boto3==1.17.74
botocore==1.20.74
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
Flask==2.0.0
idna==2.10
itsdangerous==2.0.0
Jinja2==3.0.0
jmespath==0.10.0
joblib==1.0.1
lxml==4.6.3
MarkupSafe==2.0.0
nltk==3.6.2
numpy==1.20.3
psycopg2==2.8.6
pyenchant==3.2.0
python-dateutil==2.8.1
pytorch-pretrained-bert==0.6.2
rake-nltk==1.0.4
regex==2021.4.4
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
textblob==0.15.3
torch==1.8.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
Werkzeug==2.0.0
